requirements.txt
# Core dependencies for the computer vision pipeline
torch
torchvision
scikit-learn
timm
Pillow
numpy
joblib
